Move d3d framebuffer back to media/base
